% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootsum.R
\name{bootsum}
\alias{bootsum}
\title{Visualisatie van de geschatte totale waarde van een massa financiële posten}
\usage{
bootsum(v, N, zekerheid = 0.95, b = 1e+05)
}
\arguments{
\item{v}{Een numerieke vector van positieve waarden (de getrokken steekproef).}

\item{N}{Integer. Het totale aantal posten in de massa (populatie) waaruit getrokken is.}

\item{zekerheid}{Numeric. Het zekerheidspercentage voor de minimum- en maximumgrens (standaard: 0.95).}

\item{b}{Integer. Het aantal bootstrap iteraties (standaard: 100.000).}
}
\value{
Een lijst met de geobserveerde schatting, en de minimum- en maximumgrens (invisibly).
}
\description{
Berekent en visualiseert BCa betrouwbaarheidsintervallen voor de totale
populatiewaarde op basis van een steekproef, met behulp van bootstrapping.
}
