% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_bootsum_app}
\alias{run_bootsum_app}
\title{Start de Bootsum Shiny Applicatie}
\usage{
run_bootsum_app()
}
\description{
Deze functie opent een interactieve webapplicatie in je browser
om steekproeven te evalueren met de bootsum methode.
}
